pyproject.toml
[project]
name = "hr-assit-chatbot"
version = "0.1.0"
description = "HR Assistant Chatbot MCP Server"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "mcp[cli]>=1.14.1",
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
